fix: CLI
