+++ engine/controllers/command_line_parser.cc
@@ -1,5 +1,6 @@
 #include "command_line_parser.h"
 #include "commands/chat_cmd.h"
+#include "commands/chat_completion_cmd.h"
 #include "commands/cmd_info.h"
 #include "commands/cortex_upd_cmd.h"
 #include "commands/engine_get_cmd.h"
@@ -72,7 +73,8 @@ bool CommandLineParser::SetupCommand(int argc, char** argv) {
   // Check new update, only check for stable release for now
 #ifdef CORTEX_CPP_VERSION
   if (cml_data_.check_upd) {
-    if (auto latest_version = commands::CheckNewUpdate(commands::kTimeoutCheckUpdate);
+    if (auto latest_version =
+            commands::CheckNewUpdate(commands::kTimeoutCheckUpdate);
         latest_version.has_value() && *latest_version != CORTEX_CPP_VERSION) {
       CLI_LOG("\nA new release of cortex is available: "
               << CORTEX_CPP_VERSION << " -> " << *latest_version);
@@ -108,11 +110,12 @@ void CommandLineParser::SetupCommonCommands() {
   });
 
   auto run_cmd =
-      app_.add_subcommand("run", "Shortcut to start a model and chat");
+      app_.add_subcommand("run", "Shortcut to start a model");
   run_cmd->group(kCommonCommandsGroup);
   run_cmd->usage("Usage:\n" + commands::GetCortexBinary() +
                  " run [options] [model_id]");
   run_cmd->add_option("model_id", cml_data_.model_id, "");
+  run_cmd->add_flag("--chat", cml_data_.chat_flag, "Flag for interactive mode");
   run_cmd->callback([this, run_cmd] {
     if (cml_data_.model_id.empty()) {
       CLI_LOG("[model_id] is required\n");
@@ -122,26 +125,47 @@ void CommandLineParser::SetupCommonCommands() {
     commands::RunCmd rc(cml_data_.config.apiServerHost,
                         std::stoi(cml_data_.config.apiServerPort),
                         cml_data_.model_id);
-    rc.Exec();
+    rc.Exec(cml_data_.chat_flag);
   });
 
-  auto chat_cmd = app_.add_subcommand("chat", "Send a chat completion request");
+  auto chat_cmd =
+      app_.add_subcommand("chat", "Shortcut for `cortex run --chat`");
   chat_cmd->group(kCommonCommandsGroup);
   chat_cmd->usage("Usage:\n" + commands::GetCortexBinary() +
-                  " chat [model_id] -m [msg]");
+                  " chat [model_id]");
   chat_cmd->add_option("model_id", cml_data_.model_id, "");
-  chat_cmd->add_option("-m,--message", cml_data_.msg,
-                       "Message to chat with model");
   chat_cmd->callback([this, chat_cmd] {
-    if (cml_data_.model_id.empty() || cml_data_.msg.empty()) {
-      CLI_LOG("[model_id] and [msg] are required\n");
+    if (cml_data_.model_id.empty()) {
+      CLI_LOG("[model_id] are required\n");
       CLI_LOG(chat_cmd->help());
       return;
     }
 
     commands::ChatCmd().Exec(cml_data_.config.apiServerHost,
                              std::stoi(cml_data_.config.apiServerPort),
-                             cml_data_.model_id, cml_data_.msg);
+                             cml_data_.model_id);
+  });
+
+  auto chat_completion_cmd =
+      app_.add_subcommand("chat-completion", "Send a chat completion request");
+  chat_completion_cmd->group(kCommonCommandsGroup);
+  chat_completion_cmd->usage("Usage:\n" + commands::GetCortexBinary() +
+                             " chat-complition [model_id] -m [msg]");
+  chat_completion_cmd->add_option("model_id", cml_data_.model_id, "");
+  chat_completion_cmd->add_option("-m,--message", cml_data_.msg,
+                                  "Message to chat with model");
+
+  chat_completion_cmd->callback([this, chat_completion_cmd] {
+    if (cml_data_.model_id.empty() || cml_data_.msg.empty()) {
+      CLI_LOG("[model_id] and [msg] are required\n");
+      CLI_LOG(chat_completion_cmd->help());
+      return;
+    }
+
+    commands::ChatCompletionCmd().Exec(
+        cml_data_.config.apiServerHost,
+        std::stoi(cml_data_.config.apiServerPort), cml_data_.model_id,
+        cml_data_.msg);
   });
 }
