+++ engine/controllers/command_line_parser.h
@@ -39,6 +39,7 @@ class CommandLineParser {
     std::string engine_src;
     std::string cortex_version;
     bool check_upd = true;
+    bool chat_flag = false;
     int port;
     config_yaml_utils::CortexConfig config;
     std::unordered_map<std::string, std::string> model_update_options;
